Update listing model, forms, and views
The listing model has been updated with improvements in its functionality. This includes the removal of the 'close_on' field from the model, as the status of listings is now managed by the 'isListingActive' boolean field. This makes listing status tracking more efficient. Each listing is now associated with a single 'category' instead of multiple 'categories', reflected in changes in model, forms, templates and corresponding views. The imageLink field in the ListingForm was made optional, and will now be replaced with a placeholder if not provided. The views got updated to reflect these changes in the model. A new migration file for these changes was added.