Add watchlist feature for users
Added a watchlist feature which allows users to bookmark listings they are interested in. Changes include, adding a 'watchlist' view function, 'watchlist' path in urls.py and an HTMl template for watchlist.

A many-to-many relationship was also established between the User and Listing models and a new migration file was created. Added User to the admin site to manage user instances and watchlist associations.

Updated the layout to provide a link to the watchlist page.

Improves user experience by providing a simple way to keep track of listings of interest.