WIP adding multiple forms to details
Add user authentication to listing detail view

In this commit, the "user" object is passed to the template context in the listing detail view. This is for the purpose of adding bid and comment forms as well as additional content only visible to authenticated users. If user is authenticated, they may place bids or add comments. If not, a message prompts them to log in.